Bell {multicool}	R Documentation

Compute the Bell numbers

Description

This function computes the Bell numbers, which is the summ of Stirling numbers of the second kind, S(n, k), over k, i.e.

B_n = ∑_{k=1}^{n}S(n, k),n ≥ 1

Usage

Bell(n)

B(n)

Arguments

n
A vector of one or more non-zero positive integers

Value

An vector of Bell numbers

Functions

B(): Compute the Bell numbers

Author(s)

James Curran

References

https://en.wikipedia.org/wiki/Stirling_numbers_of_the_second_kind#Recurrence_relation

See Also

Stirling2

Examples

## returns B(6)
Bell(6)

## returns B(1), B(2), ..., B(6)
B(1:6)

[Package multicool version 1.0.1 Index]