initMC {multicool}	R Documentation

Initialise the permutation object

Description

This function initialises the permutation object. It must be called before nextPerm can be called

Usage

initMC(x)

Arguments

x
a vector of integers, reals, logicals or characters

Value

a object of class mc which is a list containing elements

mode
- the mode of the original data in x, "integer", "double", or mode(x)

set
- either the multiset being permuted if mode is "integer" or a set of integers corresponding to the elements of the multiset

elements
- if mode is not "integer" then this contains the elements being permuted otherwise NULL

length
- the length of the multiset

mc
- a pointer to the internal C++ Multicool object. Users should not use this unless they really know what they are doing

Author(s)

James M. Curran

See Also

nextPerm

Examples

x = c(1,1,2,2)
m1 = initMC(x)
m1

## a non-integer example

x = rep(letters[1:4],c(2,1,2,2))
m2 = initMC(x)
m2

[Package multicool version 1.0.1 Index]